+++ .gitignore
@@ -1,4 +1,5 @@
 .ruff_cache
 .DS_store
+*.ipynb_checkpoints
 *.egg-info
 *__pycache__
