+++ src/constants.py
@@ -1,135 +1,4 @@
 ###############################################################################
-# Columns
-###############################################################################
-SYMP_COLS = [
-    "esas_pain",
-    "esas_tiredness",
-    "esas_nausea",
-    "esas_depression",
-    "esas_anxiety",
-    "esas_drowsiness",
-    "esas_appetite",
-    "esas_well_being",
-    "esas_shortness_of_breath",
-]
-LAB_COLS = [
-    "activated_partial_thromboplastin_time",
-    "alanine_aminotransferase",
-    "albumin",
-    "alkaline_phosphatase",
-    "aspartate_aminotransferase",
-    "basophil",
-    "bicarbonate",
-    "calcium",
-    "carbohydrate_antigen_19-9",
-    "chloride",
-    "creatinine",
-    "eosinophil",
-    "estimated_glomerular_filtration_rate",	
-    "glucose",
-    "hematocrit",
-    "hemoglobin",
-    "lactate_dehydrogenase",
-    "lymphocyte",
-    "magnesium",
-    "mean_corpuscular_hemoglobin",
-    "mean_corpuscular_hemoglobin_concentration",
-    "mean_corpuscular_volume",
-    "mean_platelet_volume",
-    "monocyte",
-    "neutrophil",
-    "phosphate",
-    "platelet",
-    "potassium",
-    "prothrombin_time_international_normalized_ratio",
-    "red_blood_cell",
-    "red_cell_distribution_width",
-    "sodium",
-    "total_bilirubin",
-    "white_blood_cell",
-]
-SYMP_CHANGE_COLS = [f"{col}_change" for col in SYMP_COLS]
-LAB_CHANGE_COLS = [f"{col}_change" for col in LAB_COLS]
-DRUG_COLS = [
-    '%_ideal_dose_given_GEMCITABINE',
-    '%_ideal_dose_given_CISPLATIN',
-    '%_ideal_dose_given_FLUOROURACIL',
-    '%_ideal_dose_given_ETOPOSIDE',
-    '%_ideal_dose_given_CARBOPLATIN',
-    '%_ideal_dose_given_PEMETREXED',
-    '%_ideal_dose_given_OXALIPLATIN',
-    '%_ideal_dose_given_VINORELBINE',
-    '%_ideal_dose_given_IRINOTECAN',
-    '%_ideal_dose_given_PACL',
-    '%_ideal_dose_given_NAB-PACL',
-    '%_ideal_dose_given_PEMBROLIZUMAB',
-    '%_ideal_dose_given_NIVOLUMAB',
-    '%_ideal_dose_given_DOCETAXEL',
-    '%_ideal_dose_given_EPIRUBICIN',
-    '%_ideal_dose_given_TRASTUZUMAB',
-    '%_ideal_dose_given_CYCLOPHOSPHAMIDE',
-    '%_ideal_dose_given_DOXORUBICIN',
-    '%_ideal_dose_given_DURVALUMAB',
-    '%_ideal_dose_given_BEVACIZUMAB',
-    '%_ideal_dose_given_CETUXIMAB',
-    '%_ideal_dose_given_RAMUCIRUMAB',
-    '%_ideal_dose_given_RALTITREXED',
-    '%_ideal_dose_given_TREMELIMUMAB',
-    '%_ideal_dose_given_PERTUZUMAB',
-    '%_ideal_dose_given_TOPOTECAN',
-    '%_ideal_dose_given_LIPOSOMAL IRINOTECAN',
-    '%_ideal_dose_given_ATEZOLIZUMAB',
-    '%_ideal_dose_given_IPILIMUMAB',
-    '%_ideal_dose_given_MITOMYCIN',
-    '%_ideal_dose_given_PANITUMUMAB',
-    '%_ideal_dose_given_CAPECITABINE',
-    '%_ideal_dose_given_ERLOTINIB',
-    '%_ideal_dose_given_LENVATINIB',
-    '%_ideal_dose_given_OLAPARIB',
-]
-
-# default columns for data transformations
-# imputation
-DEFAULT_IMPUTE_COLS = {
-    "mean": LAB_COLS + LAB_CHANGE_COLS,
-    "most_frequent": SYMP_COLS
-    + SYMP_CHANGE_COLS
-    + ["patient_ecog", "patient_ecog_change"],
-    "median": [],
-}
-# one-hot encoding
-DEFAULT_ENCODE_COLS = ["regimen", "intent"]
-# outlier clipping
-DEFAULT_CLIP_COLS = (
-    ["body_surface_area", "height", "weight"] + LAB_COLS + LAB_CHANGE_COLS
-)
-# normalization
-DEFAULT_NORM_COLS = (
-    [
-        "age",
-        "body_surface_area",
-        "cycle_number",
-        "days_since_last_treatment",
-        "days_since_prev_ED_visit",
-        "days_since_starting_treatment",
-        "height",
-        "line_of_therapy",
-        "num_prior_ED_visits_within_5_years",
-        "patient_ecog",
-        "patient_ecog_change",
-        "visit_month_sin",
-        "visit_month_cos",
-        "weight",
-    ]
-    + DRUG_COLS
-    + LAB_COLS
-    + LAB_CHANGE_COLS
-    + SYMP_COLS
-    + SYMP_CHANGE_COLS
-)
-
-
-###############################################################################
 # Mappings
 ###############################################################################
 # cancer site / cancer morphology
@@ -286,47 +155,4 @@ CANCER_CODE_MAP = {
     "947": "Gliomas",
     "954": "Nerve sheath tumors",
     "956": "Nerve sheath tumors",
-}
-
-# units
-UNIT_MAP = {
-    "L/L": ["hematocrit"],
-    "U/L": [
-        "alanine_aminotransferase",
-        "alkaline_phosphatase",
-        "aspartate_aminotransferase",
-        "lactate_dehydrogenase",
-    ],
-    "cm": ["height"],
-    "fL": ["mean_corpuscular_volume", "mean_platelet_volume"],
-    "g/L": ["albumin", "hemoglobin", "mean_corpuscular_hemoglobin_concentration"],
-    "kg": ["weight"],
-    "kU/L": ["carbohydrate_antigen_19-9"],
-    "mL/min/1.73m2": ["eGFR"],
-    "mmol/L": [
-        "bicarbonate",
-        "calcium",
-        "chloride",
-        "glucose",
-        "magnesium",
-        "phosphate",
-        "potassium",
-        "sodium",
-    ],
-    "m^2": ["body_surface_area"],
-    "pg": ["mean_corpuscular_hemoglobin"],
-    "s": ["activated_partial_thromboplastin_time"],
-    "umol/L": ["creatinine", "total_bilirubin"],
-    "x10e9/L": [
-        "basophil",
-        "eosinophil",
-        "lymphocyte",
-        "monocyte",
-        "neutrophil",
-        "platelet",
-        "white_blood_cell",
-    ],
-    "x10e12/L": ["red_blood_cell"],
-    "years": ["age"],
-    "%CV": ["red_cell_distribution_width"],
 }
