+++ src/util.py
@@ -1,7 +1,7 @@
-import os
 import itertools
 import logging
 import multiprocessing as mp
+import os
 import pickle
 
 import numpy as np
@@ -69,42 +69,17 @@ def split_and_parallelize(
     """
     if split_by_mrns:
         generator = []
-        mrns = data[0]["mrn"] if isinstance(data, tuple) else data["mrn"]
-        mrn_groupings = np.array_split(mrns.unique(), processes)
         if isinstance(data, tuple):
+            mrn_groupings = np.array_split(data[0]["mrn"].unique(), processes)
             for mrn_grouping in mrn_groupings:
                 items = tuple(df[df["mrn"].isin(mrn_grouping)] for df in data)
                 generator.append(items)
         else:
+            mrn_groupings = np.array_split(data["mrn"].unique(), processes)
             for mrn_grouping in mrn_groupings:
-                item = data[mrns.isin(mrn_grouping)]
+                item = data[data["mrn"].isin(mrn_grouping)]
                 generator.append(item)
     else:
         # splits df into x number of partitions, where x is number of processes
         generator = np.array_split(data, processes)
     return parallelize(generator, worker, processes=processes)
-
-
-###############################################################################
-# Data Descriptions
-###############################################################################
-def get_nunique_categories(df: pd.DataFrame) -> pd.DataFrame:
-    catcols = df.dtypes[df.dtypes == object].index.tolist()
-    return pd.DataFrame(
-        df[catcols].nunique(), columns=["Number of Unique Categories"]
-    ).T
-
-
-def get_nmissing(df: pd.DataFrame) -> pd.DataFrame:
-    missing = df.isnull().sum()  # number of nans for each column
-    missing = missing[missing != 0]  # remove columns without missing values
-    missing = pd.DataFrame(missing, columns=["Missing (N)"])
-    missing["Missing (%)"] = (missing["Missing (N)"] / len(df) * 100).round(3)
-    return missing.sort_values(by="Missing (N)")
-
-
-def get_excluded_numbers(df, mask: pd.Series, context: str = ".") -> None:
-    """Report the number of patients and sessions that were excluded"""
-    N_sessions = sum(~mask)
-    N_patients = len(set(df["mrn"]) - set(df.loc[mask, "mrn"]))
-    logger.info(f"Removing {N_patients} patients and {N_sessions} sessions{context}")
